service to find user by id and email where made in microservice-user

+++ microservice-user/src/main/java/com/service/microservice_user/services/UserService.java
@@ -8,10 +8,10 @@ import java.util.List;
 public interface UserService {
 
     List<UserEntity> getAllUsers();
-    UserDTO getUserById(int id);
+    UserDTO getUserById(long id);
     UserDTO getUserByEmail(String email);
     int createUser(UserDTO user);
     int updateUser(UserDTO user);
-    int deleteUser(int id);
+    int deleteUser(long id);
 
 }
